[FIX/#71] SetTagView 리팩토링
